fixed the error bezierplot "sin(8*x)" 0 "4*pi"
The problem was that somehow inf instead of a number was part of the bezierpath. I simply added a check that such a line is not added to the output. Maybe later I will add (again) a function that detects sinusoid functions, which will fasten up things quite much and improve the output (fewer points).