got rid of debug (lualatex 1.18 no longer loads it)
Thelua script has to work as a "main" with an external call from pdfLaTeX as well as a "library" with a direct Lua use of the function bezierplot() when used with LuaLaTeX. However, Lua does not seem to support a natural distinction between the main and the library (https://stackoverflow.com/questions/4521085/main-function-in-lua),. As https://www.tug.org/pipermail/luatex/2024-February/007935.html states, LuaTeX 1.18 has disabled the use of the debug library. Hence, using "pcall(debug.getlocal, 4, 1)" is no longer possible without restriction. The problem with the now chosen method is that arg could be defined as a global variable in the calling code.
The functions cbrt() and sgn() are now local.

The style file has only a new date (to match the lua script date).